backend/requirements.txt
alembic==1.13.2
sqlalchemy==2.0.31
fastapi>=0.45.0,<0.46.0
beautifulsoup4==4.12.3
lxml==5.2.2
bcrypt==4.1.3
typer[all]
rich==13.7.0
click==8.1.7
click-completion==0.5.2
requests==2.32.0
python-decouple==3.8
psycopg2==2.9.9
uvicorn==0.30.1
pydantic[email]
PyJWT==2.7.0